refactor(base_transformers): move module and datamodule to modules dir
